Move specific functions from src/lib/ to src/utils/
