+++ src/components/Tags.js
@@ -4,8 +4,8 @@ import React, { Component } from 'react'
 import Category from './Category'
 import More from '../components/More'
 import _ from 'lodash'
-import { imageComposer } from '../lib/image-composer.js'
 import { date2yyyymmdd } from '../lib/date-transformer'
+import { imageComposer } from '../utils/index'
 
 if (process.env.BROWSER) {
   require('./Tags.css')
