+++ src/components/article/Image.js
@@ -4,7 +4,7 @@ import _ from 'lodash'
 import classNames from 'classnames'
 import commonStyles from './Common.scss'
 import FitwidthMixin from './mixins/FitwidthMixin'
-import { getScreenType } from '../../lib/screen-type'
+import { getScreenType } from '../../utils/index'
 import BlockAlignmentWrapper from './BlockAlignmentWrapper'
 import React, { Component } from 'react'
 import ReactDOM from 'react-dom'
