+++ src/components/article/ImageDiff.js
@@ -4,7 +4,7 @@ import _ from 'lodash'
 import BlockAlignmentWrapper from './BlockAlignmentWrapper'
 import classNames from 'classnames'
 import FitwidthMixin from './mixins/FitwidthMixin'
-import { getScreenType } from '../../lib/screen-type'
+import { getScreenType } from '../../utils/index'
 import MediaQuery from 'react-responsive'
 import React, { Component } from 'react'
 import ReactDOM from 'react-dom'
