+++ src/utils/detect-device.js
