+++ src/utils/get-screen-type.js
