+++ src/utils/image-composer.js
