+++ src/utils/index.js
@@ -1,4 +1,8 @@
 'use strict'
 export * from './formaturl'
 export * from './denormalize-articles'
-export * from './article-nested-entity-methods'
+export * from './detect-device'
+export * from './image-composer'
+export * from './get-embedded-query-param'
+export * from './get-screen-type'
+export * from './url-transformer'
